--- Bookdata/book.json
@@ -1,11 +1,0 @@
-[
-  {
-    "id": 702,
-    "title": "I hop",
-    "publisher": "good",
-    "title_label": "Karla",
-    "color": "Blue",
-    "cover_state": "03/10/2018",
-    "publish_date": "It is working"
-  }
-]
