--- Bookdata/label.json
@@ -1,7 +1,0 @@
-[
-  {
-    "id": 151,
-    "title": "Karla",
-    "color": "Blue"
-  }
-]
