+++ json/book.json
@@ -1,0 +1,11 @@
+[
+  {
+    "id": 750,
+    "title": "why",
+    "publisher": "good",
+    "title_label": "hello",
+    "color": "red",
+    "cover_state": "09/10/2010",
+    "publish_date": "it is not working"
+  }
+]
