+++ json/label.json
@@ -1,0 +1,7 @@
+[
+  {
+    "id": 92,
+    "title": "hello",
+    "color": "red"
+  }
+]
